requirements.txt
fastapi
uvicorn
xgboost
scikit-learn
pandas
numpy
joblib
pandera
mlflow
optuna
s3fs
prefect
flake8
scipy
lightgbm
catboost
shap
lime
umap-learn
imbalanced-learn
